# Cache flops_per_fwdbwd in estimate_mfu