Minor performance improvement for the `estimate_mfu` function.

## Profiling results:

### Before
```
2948092 function calls (2099792 primitive calls) in 0.693 seconds

   Ordered by: standard name

   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
   225548    0.028    0.000    0.042    0.000 _tensor.py:1002(__hash__)
3992/1996    0.004    0.000    0.005    0.000 eval_frame.py:149(__getattr__)
     1996    0.006    0.000    0.688    0.000 model.py:297(estimate_mfu)
   113772    0.037    0.000    0.670    0.000 model.py:301(<genexpr>)
   113772    0.139    0.000    0.575    0.000 module.py:2156(_named_members)
   113772    0.023    0.000    0.619    0.000 module.py:2170(parameters)
   113772    0.021    0.000    0.595    0.000 module.py:2195(named_parameters)
   191616    0.044    0.000    0.055    0.000 module.py:2224(<lambda>)
1039916/193612    0.281    0.000    0.320    0.000 module.py:2335(named_modules)
     1996    0.001    0.000    0.001    0.000 {built-in method builtins.getattr}
   225548    0.014    0.000    0.014    0.000 {built-in method builtins.id}
     1996    0.013    0.000    0.682    0.000 {built-in method builtins.sum}
   303392    0.047    0.000    0.065    0.000 {method 'add' of 'set' objects}
     1996    0.000    0.000    0.000    0.000 {method 'disable' of '_lsprof.Profiler' objects}
   383232    0.022    0.000    0.022    0.000 {method 'items' of 'collections.OrderedDict' objects}
   111776    0.014    0.000    0.014    0.000 {method 'numel' of 'torch._C._TensorBase' objects}
```
### After
```
11461 function calls (9040 primitive calls) in 0.007 seconds

   Ordered by: standard name

   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
      113    0.000    0.000    0.000    0.000 _tensor.py:1002(__hash__)
3992/1996    0.004    0.000    0.004    0.000 eval_frame.py:149(__getattr__)
     1996    0.002    0.000    0.003    0.000 model.py:298(estimate_mfu)
       57    0.000    0.000    0.000    0.000 model.py:303(<genexpr>)
        1    0.000    0.000    0.000    0.000 module.py:1697(__setattr__)
       57    0.000    0.000    0.000    0.000 module.py:2156(_named_members)
       57    0.000    0.000    0.000    0.000 module.py:2170(parameters)
       57    0.000    0.000    0.000    0.000 module.py:2195(named_parameters)
       96    0.000    0.000    0.000    0.000 module.py:2224(<lambda>)
   521/97    0.000    0.000    0.000    0.000 module.py:2335(named_modules)
        1    0.000    0.000    0.000    0.000 parameter.py:8(__instancecheck__)
     1996    0.001    0.000    0.001    0.000 {built-in method builtins.getattr}
      113    0.000    0.000    0.000    0.000 {built-in method builtins.id}
      3/2    0.000    0.000    0.000    0.000 {built-in method builtins.isinstance}
        1    0.000    0.000    0.000    0.000 {built-in method builtins.sum}
        1    0.000    0.000    0.000    0.000 {function _ParameterMeta.__instancecheck__ at 0x7ff934943370}
      152    0.000    0.000    0.000    0.000 {method 'add' of 'set' objects}
     1996    0.000    0.000    0.000    0.000 {method 'disable' of '_lsprof.Profiler' objects}
        3    0.000    0.000    0.000    0.000 {method 'get' of 'dict' objects}
      192    0.000    0.000    0.000    0.000 {method 'items' of 'collections.OrderedDict' objects}
       56    0.000    0.000    0.000    0.000 {method 'numel' of 'torch._C._TensorBase' objects}
 ```