@karpathy 2024-02-13
this can't possibly be taking too long and bloats code reject