@GilesBathgate 2024-02-13
Agreed that it's not significant. But disagree that it's bloat in the context of the mfu calculation which itself is bloat. ;)